Update pokemonAdder.pde
Works with new info format now (first field is string array of [new pokemon name, trues and falses in order of questions, new pokemon type] in the order that it needs to be written to the file in, second field is just the new question string to be added. Take a new question every time a new pokemon is added, because when an empty new question field is passed, it will still write new trues and falses to the file).